Updated tblBenSamps to have a fk to tlkCommunities. Updated the Communities editor to handle deleting already asigned samples and setting messages.

+++ RWInbound2/Edit/EditCommunities.aspx.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Data.Entity.Validation;
 using System.Linq;
 using System.Reflection;
@@ -16,16 +17,34 @@ namespace RWInbound2.Edit
         {
             if (!IsPostBack)
             {
+                SetMessages();
                 // Validate initially to force asterisks
                 // to appear before the first roundtrip.
                 Validate();
             }
-            ErrorLabel.Text = "";
-            SuccessLabel.Text = "";
+        }
+
+        private void SetMessages(string type = "", string message = "")
+        {
+            switch (type)
+            {
+                case "Success":
+                    ErrorLabel.Text = "";
+                    SuccessLabel.Text = message;
+                    break;
+                case "Error":
+                    ErrorLabel.Text = message;
+                    SuccessLabel.Text = "";
+                    break;
+                default:
+                    ErrorLabel.Text = "";
+                    SuccessLabel.Text = "";
+                    break;
+            }
         }
 
         public IQueryable<tlkCommunity> GetCommunities([QueryString]string descriptionSearchTerm = "",
-                                                                     [QueryString]string successLabelMessage = "")
+                                                       [QueryString]string successLabelMessage = "")
         {
             try
             {
@@ -33,8 +52,7 @@ namespace RWInbound2.Edit
 
                 if (!string.IsNullOrEmpty(successLabelMessage))
                 {
-                    ErrorLabel.Text = "";
-                    SuccessLabel.Text = successLabelMessage;
+                    SetMessages("Success", successLabelMessage);
                 }
 
                 if (!string.IsNullOrEmpty(descriptionSearchTerm))
@@ -117,6 +135,8 @@ namespace RWInbound2.Edit
         {
             try
             {
+                SetMessages();
+
                 using (RiverWatchEntities _db = new RiverWatchEntities())
                 {
                     var communityToUpdate = _db.tlkCommunities.Find(model.ID);
@@ -137,8 +157,8 @@ namespace RWInbound2.Edit
                     communityToUpdate.DateLastModified = DateTime.Now;
                     _db.SaveChanges();
 
-                    ErrorLabel.Text = "";
-                    SuccessLabel.Text = "Community Updated";
+                    string successMsg = string.Format("Community Updated: {0}", model.Description);
+                    SetMessages("Success", successMsg);
                 }
             }
             catch (Exception ex)
@@ -153,11 +173,30 @@ namespace RWInbound2.Edit
             {
                 try
                 {
-                    var communityToDelete = _db.tlkCommunities.Find(model.ID);
-                    _db.tlkCommunities.Remove(communityToDelete);
-                    _db.SaveChanges();
-                    ErrorLabel.Text = "";
-                    SuccessLabel.Text = "Community Deleted";
+                    SetMessages();
+
+                    var tblBenSampSampleNumbers = (from bs in _db.tblBenSamps
+                                                   join s in _db.Samples on bs.SampleID equals s.ID
+                                                   where bs.Community == model.ID
+                                                   orderby s.SampleNumber
+                                                   select s.SampleNumber).ToList();
+
+                    if (tblBenSampSampleNumbers.Count > 0)
+                    {
+                        string errorMsg
+                            = string.Format("Community {1} can not be deleted because it is assigned to one or more Benthic Samples. {0}",
+                                            "\r\n Click the Download Assigned Samples button to get a list of the samples.", model.Description);
+                        SetMessages("Error", errorMsg);
+                    }
+                    else
+                    {
+                        var communityToDelete = _db.tlkCommunities.Find(model.ID);
+                        _db.tlkCommunities.Remove(communityToDelete);
+                        _db.SaveChanges();
+
+                        string successMsg = string.Format("Community Deleted: {0}", model.Description);
+                        SetMessages("Success", successMsg);
+                    }                    
                 }
                 catch (Exception ex)
                 {
@@ -170,13 +209,13 @@ namespace RWInbound2.Edit
         {
             try
             {
+                SetMessages();
                 string strCode = ((TextBox)CommunitiesGridView.FooterRow.FindControl("NewCode")).Text;
                 string description = ((TextBox)CommunitiesGridView.FooterRow.FindControl("NewDescription")).Text;
 
                 if (string.IsNullOrEmpty(strCode))
                 {
-                    SuccessLabel.Text = "";
-                    ErrorLabel.Text = "Code field is required.";
+                    SetMessages("Error", "Code field is required.");
                 }
                 else
                 {
@@ -184,8 +223,7 @@ namespace RWInbound2.Edit
                     bool convertToInt = int.TryParse(strCode, out code);
                     if (!convertToInt)
                     {
-                        SuccessLabel.Text = "";
-                        ErrorLabel.Text = "Code field must be an integer number.";
+                        SetMessages("Error", "Code field must be an integer number.");
                     }
                     else
                     {
@@ -227,6 +265,62 @@ namespace RWInbound2.Edit
             }
         }
 
+        protected void CommunitiesGridView_RowCommand(object sender, GridViewCommandEventArgs e)
+        {
+            try
+            {
+                if (e.CommandName == "GetAssignedSamples")
+                {
+                    using (RiverWatchEntities _db = new RiverWatchEntities())
+                    {
+                        if (e.CommandName == "GetAssignedSamples")
+                        {
+                            int iD = Convert.ToInt32(e.CommandArgument);
+
+                            // Retrieve the row that contains the button 
+                            // from the Rows collection.
+                            //GridViewRow row = FieldProceduresGridView.Rows[iD];
+
+                            var tblBenSampSampleNumbers = (from bs in _db.tblBenSamps
+                                                           join s in _db.Samples on bs.SampleID equals s.ID
+                                                           where bs.Community == iD
+                                                           orderby s.SampleNumber
+                                                           select s.SampleNumber).Distinct().ToList();
+
+                            if (tblBenSampSampleNumbers.Count > 0)
+                            {
+                                //return csv to show Assigned Sample numbers
+                                Response.ClearContent();
+                                Response.AddHeader("content-disposition", string.Format("attachment; filename={0}", "Assigned_SampleNumbers.csv"));
+                                Response.ContentType = "application/text";
+                                string csvContents = GetSamplesCSVData(tblBenSampSampleNumbers);
+                                Response.Write(csvContents);
+                                Response.End();
+                            }
+                            else
+                            {
+                                SetMessages("Error", "There are 0 Benthic Samples assigned to this item.");
+                            }
+                        }
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                HandleErrors(ex, ex.Message, "CommunitiesGridView_RowCommand", "", "");
+            }
+        }
+
+        private string GetSamplesCSVData(List<string> listOfSamples)
+        {
+            StringBuilder strbldr = new StringBuilder();
+            foreach (var sample in listOfSamples)
+            {
+                strbldr.AppendLine(sample);
+            }
+            return strbldr.ToString();
+        }
+
         private void HandleErrors(Exception ex, string msg, string fromPage,
                                                 string nam, string comment)
         {
@@ -257,6 +351,11 @@ namespace RWInbound2.Edit
                 SuccessLabel.Text = "";
                 ErrorLabel.Text = ex.Message;
             }
+        }
+
+        protected void CommunitiesGridView_RowEditing(object sender, GridViewEditEventArgs e)
+        {
+            SetMessages();
         }
     }
 }
